log_management/requirements.txt
amqp==2.4.1
urllib3==1.24.2
asn1crypto==0.24.0
backports-abc==0.5
backports.ssl-match-hostname==3.7.0.1
billiard==3.5.0.5
cachetools==3.1.1
celery==4.2.1
certifi==2018.11.29
cffi==1.9.1
chardet==3.0.4
cryptography==2.3
future==0.17.1
futures==3.1.1
google-auth==1.9.0
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
kombu==4.3.0
kubernetes==10.0.1
numpy==1.16.1
oauthlib==3.1.0
pandas==0.24.1
ply==3.9
protobuf==3.11.2
psycopg2-binary==2.7.7
pyarrow==0.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.14
pyOpenSSL==17.3.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.2
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
tornado==4.4.2
tornado-cors==0.6.0
vine==1.2.0
virtualenv==16.7.9
websocket-client==0.56.0
